move batch_process_answer to answer_operation
code clean up and move batch_process_answer and its test function to answer_operation
