+++ back-end/www/tests/location_tests.py
@@ -177,177 +177,6 @@ class LocationTest(BasicTest):
         count = location_operations.get_location_count()
         assert count == 4
 
-    def test_batch_process_answers(self):
-        """
-        Location #l1 has gold standard.
-        User u1 passes the standard test, and submit answers to #l2 and #l3.
-        User u2 fails the standard test, only #l2 matches u1's answer.
-        User u3 passes the standard test, only #l2 matches u1's answer.
-        User u4 passes the standard test, only #l4 matches u1's answer.
-        Pass if location done_at correct after each user's answer submit, and individual_done_count correct.
-        """
-        IS_GOLD_STANDARD = 0
-        user1 = user_operations.create_user("123")
-        user2 = user_operations.create_user("456")
-        user3 = user_operations.create_user("789")
-        user_admin = user_operations.create_user("ADMIN")
-        l1 = location_operations.create_location("AAA")
-        l2 = location_operations.create_location("BBB")
-        l3 = location_operations.create_location("CCC")
-
-        assert(l2.done_at==None)
-
-        # l1 has gold answer.
-        A_gold = answer_operations.create_answer(user_admin.id, l1.id, 2000, 2010, "", 1, 1, IS_GOLD_STANDARD,
-                0, 0, 0, 0, 0)
-
-        user1_answers=[
-            {"location_id": l1.id,
-             "year_new": 2000,
-             "year_old": 1997,
-             "zoom_level": 0,
-             "left_top_lat": 0,
-             "left_top_lng": 0,
-             "bbox_left_top_lat": 0,
-             "bbox_left_top_lng": 0,
-             "bbox_bottom_right_lat": 0,
-             "bbox_bottom_right_lng": 0,
-             "land_usage": 1,
-             "expansion": 1,
-             "source_url_root": "xxx"},
-            {"location_id": l2.id,
-             "year_new": 2000,
-             "year_old": 1997,
-             "zoom_level": 0,
-             "left_top_lat": 0,
-             "left_top_lng": 0,
-             "bbox_left_top_lat": 0,
-             "bbox_left_top_lng": 0,
-             "bbox_bottom_right_lat": 0,
-             "bbox_bottom_right_lng": 0,
-             "land_usage": 1,
-             "expansion": 1,
-             "source_url_root": "xxx"},
-            {"location_id": l3.id,
-             "year_new": 2000,
-             "year_old": 1997,
-             "zoom_level": 0,
-             "left_top_lat": 0,
-             "left_top_lng": 0,
-             "bbox_left_top_lat": 0,
-             "bbox_left_top_lng": 0,
-             "bbox_bottom_right_lat": 0,
-             "bbox_bottom_right_lng": 0,
-             "land_usage": 1,
-             "expansion": 1,
-             "source_url_root": "xxx"},
-        ]
-
-        user2_answers=[
-            {"location_id": l1.id,
-             "year_new": 2000,
-             "year_old": 1997,
-             "zoom_level": 0,
-             "left_top_lat": 0,
-             "left_top_lng": 0,
-             "bbox_left_top_lat": 0,
-             "bbox_left_top_lng": 0,
-             "bbox_bottom_right_lat": 0,
-             "bbox_bottom_right_lng": 0,
-             "land_usage": 0,
-             "expansion": 0,
-             "source_url_root": "xxx"},
-            {"location_id": l2.id,
-             "year_new": 2000,
-             "year_old": 1997,
-             "zoom_level": 0,
-             "left_top_lat": 0,
-             "left_top_lng": 0,
-             "bbox_left_top_lat": 0,
-             "bbox_left_top_lng": 0,
-             "bbox_bottom_right_lat": 0,
-             "bbox_bottom_right_lng": 0,
-             "land_usage": 1,
-             "expansion": 1,
-             "source_url_root": "xxx"},
-            {"location_id": l3.id,
-             "year_new": 2000,
-             "year_old": 1997,
-             "zoom_level": 0,
-             "left_top_lat": 0,
-             "left_top_lng": 0,
-             "bbox_left_top_lat": 0,
-             "bbox_left_top_lng": 0,
-             "bbox_bottom_right_lat": 0,
-             "bbox_bottom_right_lng": 0,
-             "land_usage": 0,
-             "expansion": 0,
-             "source_url_root": "xxx"},
-        ]
-
-        user3_answers=[
-            {"location_id": l1.id,
-             "year_new": 2000,
-             "year_old": 1997,
-             "zoom_level": 0,
-             "left_top_lat": 0,
-             "left_top_lng": 0,
-             "bbox_left_top_lat": 0,
-             "bbox_left_top_lng": 0,
-             "bbox_bottom_right_lat": 0,
-             "bbox_bottom_right_lng": 0,
-             "land_usage": 1,
-             "expansion": 1,
-             "source_url_root": "xxx"},
-            {"location_id": l2.id,
-             "year_new": 2000,
-             "year_old": 1997,
-             "zoom_level": 0,
-             "left_top_lat": 0,
-             "left_top_lng": 0,
-             "bbox_left_top_lat": 0,
-             "bbox_left_top_lng": 0,
-             "bbox_bottom_right_lat": 0,
-             "bbox_bottom_right_lng": 0,
-             "land_usage": 1,
-             "expansion": 1,
-             "source_url_root": "xxx"},
-            {"location_id": l3.id,
-             "year_new": 2000,
-             "year_old": 1997,
-             "zoom_level": 0,
-             "left_top_lat": 0,
-             "left_top_lng": 0,
-             "bbox_left_top_lat": 0,
-             "bbox_left_top_lng": 0,
-             "bbox_bottom_right_lat": 0,
-             "bbox_bottom_right_lng": 0,
-             "land_usage": 0,
-             "expansion": 0,
-             "source_url_root": "xxx"},
-        ]
-
-        # User u1 passes the standard test, and submit answers to #l2 and #l3.
-        result = location_operations.batch_process_answers(user1.id, user1_answers)
-        assert(result==True)
-        assert(l2.done_at is None)
-        assert(l3.done_at is None)
-
-        # User u2 fails the standard test, only #l2 matches u1.
-        result = location_operations.batch_process_answers(user2.id, user2_answers)
-        assert(result==False)
-        assert(l2.done_at is None)
-        assert(l3.done_at is None)
-
-        # User u3 passes the standard test, only #l2 matches u1. #l2 Done criteria reaches.
-        result = location_operations.batch_process_answers(user3.id, user3_answers)
-        assert(result==True)
-        assert(l2.done_at is not None)
-        assert(l3.done_at is None)
-
-        loc_count = user_operations.get_user_done_location_count(user2.id)
-        assert(loc_count == len(user2_answers))
-
 
 if __name__ == "__main__":
     unittest.main()
